requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.0.6
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-multiselectfield2==0.0.1
django-widget-tweaks==1.4.12
flake8==5.0.4
future==0.18.2
gunicorn==20.1.0
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.1
regex==2022.10.31
schedule==1.1.0
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.1
whitenoise==6.2.0
